requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cleanup==8.0.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.2
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
fontawesomefree==6.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
